requirements.txt
pandas
faker
pinecone-client>=2.2.1
openai>=0.27.0
python-dotenv>=1.0.0